Move lnaguage as C into codegeneration
