+++ ffc/codegeneration/C/cnodes.py
@@ -10,9 +10,9 @@ import numbers
 
 import numpy
 
-from ffc.language.format_lines import Indented, format_indented_lines
-from ffc.language.format_value import (format_float, format_int, format_value)
-from ffc.language.precedence import PRECEDENCE
+from ffc.codegeneration.C.format_lines import Indented, format_indented_lines
+from ffc.codegeneration.C.format_value import (format_float, format_int, format_value)
+from ffc.codegeneration.C.precedence import PRECEDENCE
 
 logger = logging.getLogger(__name__)
 """CNode TODO:
