+++ setup.py
@@ -123,7 +123,6 @@ def run_install():
         packages=[
             "ffc",
             "ffc.codegeneration",
-            "ffc.language",
             "ffc.tsfc",
             "ffc.uflacs",
             "ffc.uflacs.analysis",
